Move files to Base and createWorkflow folders
