+++ frontend/src/components/Base/AppBar.jsx
@@ -8,7 +8,7 @@ import {
 	IconButton,
 } from "@mui/material";
 import HowToVoteIcon from "@mui/icons-material/HowToVote";
-import MUISwitch from "components/MuiSwitch";
+import MUISwitch from "components/Base/MuiSwitch";
 
 const TopAppBar = ({ dark, setDark }) => {
 	const toggleSwitch = () => {
