+++ frontend/src/components/createWorkflow/ReviewDetails.jsx
@@ -5,8 +5,8 @@ import MuiAccordion from "@mui/material/Accordion";
 import MuiAccordionSummary from "@mui/material/AccordionSummary";
 import MuiAccordionDetails from "@mui/material/AccordionDetails";
 import { styled } from "@mui/system";
-import StickyHeadTable from "components/RenderTable";
-import TimesTable from "components/TimesTable";
+import ReviewInfoTable from "components/Base/RenderTable";
+import TimesTable from "components/createWorkflow/TimesTable";
 
 const Accordion = styled((props) => (
 	<MuiAccordion disableGutters elevation={0} square {...props} />
@@ -85,7 +85,7 @@ const ReviewDetails = () => {
 							maxHeight: "calc(80vh - 200px)",
 						}}
 					>
-						<StickyHeadTable stage="participants" />
+						<ReviewInfoTable stage="participants" />
 					</div>
 				</AccordionDetails>
 			</Accordion>
@@ -106,7 +106,7 @@ const ReviewDetails = () => {
 							maxHeight: "calc(80vh - 200px)",
 						}}
 					>
-						<StickyHeadTable stage="candidates" />
+						<ReviewInfoTable stage="candidates" />
 					</div>
 				</AccordionDetails>
 			</Accordion>
