+++ frontend/src/components/createWorkflow/SelectParticipants.jsx
@@ -1,7 +1,7 @@
 import React, { useContext } from "react";
 import { Button, Typography, ButtonGroup, TextField } from "@mui/material";
 import { VoteInfoContext } from "context/VoteInfoContext";
-import StickyHeadTable from "components/RenderTable";
+import StickyHeadTable from "components/Base/RenderTable";
 import { txtUploadHandler, excelUploadHandler } from "utils/FileUpload";
 import { cancelVote } from "utils/CancelVote";
 import { generateAlgorandAccounts } from "utils/AlgoFunctions";
@@ -14,9 +14,9 @@ import {
 	ManualUploadDiv,
 	ManualUploadSubDiv,
 } from "utils/ParticipantsStyle";
-import CustomizedInputBase from "components/TextInput";
+import CustomizedInputBase from "components/createWorkflow/TextInput";
 import HelpIcon from "@mui/icons-material/Help";
-import HelperTooltip from "components/HelperTooltip";
+import HelperTooltip from "components/createWorkflow/HelperTooltip";
 
 // const emails1 = require("images/emails1.png");
 // const emails2 = require("images/emails2.png");
